Ai_research_Summarizer/requirements.txt
arxiv
transformers
fpdf
python-dotenv
